Extract `FeedImageDataStoreSpy` into a shared scope
